Reorganize indexes, work on docs
