+++ dynamorm/__init__.py
@@ -5,5 +5,7 @@
     from dynamorm import DynaModel
 
 """
-from .model import DynaModel, GlobalIndex, LocalIndex, ProjectAll, ProjectKeys, ProjectInclude  # noqa
+from .model import DynaModel  # noqa
+from .indexes import GlobalIndex, LocalIndex, ProjectAll, ProjectKeys, ProjectInclude  # noqa
+from .relationships import ManyToOne, OneToMany, OneToOne  # noqa
 from .table import Q  # noqa
